+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+Changed:
+
+* Changed canonical English name from Chattogram to Chittagong for compatibility.
+* Renamed division folder from data/divisions/chattogram to data/divisions/chittagong.
+
 ## v0.1.0-barishal
 
 Released: Initial public pre-release